refactor: split markdown focus boundary
